requirements.txt
biopython
pyyaml
whisperx
torchaudio
